Shit son what are you doing
I've moved a shader into object. Some small issues with that currently (models color is now black instead of white)
I've created an object manager which stores all the objects in the scene. I intend to store a pointer to the camera projection view matrix so I don't need to pass it in.  It is being passed in but not being updated so it isn't pointing to the location.