Added light class
Intended to store all lights created in scene. Need a specific point to create the light.
Note that you can pass in arrays of lights therefore doing all the math for the lights and having an object lit from multiple sides.
Need to think about how a point light will do it's thing.