I regret this
I've added a bunch of managers and finished separating the object creation process. I now need an asset manager that will be in charge of loading the assets at app start. Then a way for them to be looked for and applied to objects. Also don't forget fix the draw loop to add the other texturemaps you're looking for. That is a priority. Needs to be done by the week.